Good stopping place.
Can now extract bunch of info of a verse from biblehub, also have whole bible structure. Next up is to make it do concurrent requests so it doesn't take forever to extract the whole bible. I am still not 100% sure if the structure of the bible is correct, like there may be a few verses missing or extra.

Next is to extract the whole biblehub. Save it in a sqllite database probably. And then maybe scrape more info from biblehub. Maybe develop an API I can use to query data easier.